uint to size_t

+++ src/stage/dacc.cc
@@ -23,13 +23,13 @@ namespace rdb
         myFile.close();
     };
 
-    void DataAccessor::Update(const std::byte *outBuffer, const uint offsetFromHead)
+    void DataAccessor::Update(const std::byte *outBuffer, const size_t offsetFromHead)
     {
         auto size = descriptor.GetSize();
         pAccessor->Update(outBuffer, size, offsetFromHead * size);
     };
 
-    void DataAccessor::Get(std::byte *inBuffer, const uint offsetFromHead)
+    void DataAccessor::Get(std::byte *inBuffer, const size_t offsetFromHead)
     {
         auto size = descriptor.GetSize();
         pAccessor->Read(inBuffer, size, offsetFromHead * size);
